:recycle: retire the `drmem-config` crate
All the code in the `drmem-config` crate has been moved closer to the
modules that use the configuration types. The crate was getting too
complicated with the feature conditions. This should clen most of that
up.
