+++ backends/drmem-db-redis/Cargo.toml
@@ -14,6 +14,9 @@ keywords = ["control-system", "automation", "database"]
 maintenance = { status = "actively-developed" }
 
 [dependencies]
+toml = "0.5.6"
+serde = "1.0.115"
+serde_derive = "1.0.115"
 async-trait = "0.1"
 redis = { version = "0.22", features = ["tokio-comp"] }
 tracing-subscriber = "^0.2"
@@ -24,8 +27,3 @@ futures = { version = "0.3", features = ["std"] }
 futures-util = "0.3"
 
 drmem-api = { path = "../../drmem-api", version = "0.1" }
-
-[dependencies.drmem-config]
-path = "../../drmem-config"
-version = "0.1"
-features = ["redis-backend"]
