+++ backends/drmem-db-simple/Cargo.toml
@@ -14,12 +14,12 @@ keywords = ["control-system", "automation"]
 maintenance = { status = "actively-developed" }
 
 [dependencies]
+toml = "0.5.6"
+serde = "1.0.115"
+serde_derive = "1.0.115"
 async-trait = "0.1"
-drmem-api = { path = "../../drmem-api", version = "0.1" }
 tokio = { version = "1", features = ["sync"] }
 tokio-stream = { version = "0.1", features = ["sync"] }
 tracing = { version = "^0.1", features = ["std"] }
 
-[dependencies.drmem-config]
-path = "../../drmem-config"
-version = "0.1"
+drmem-api = { path = "../../drmem-api", version = "0.1" }
