+++ backends/drmem-db-simple/src/config.rs
@@ -1,0 +1,12 @@
+use serde_derive::Deserialize;
+
+#[derive(Deserialize, Clone)]
+pub struct Config {}
+
+impl Config {
+    pub const fn new() -> Config {
+        Config {}
+    }
+}
+
+pub static DEF: Config = Config::new();
