+++ drivers/drmem-drv-weather-wu/Cargo.toml
@@ -17,6 +17,5 @@ tracing-futures = "^0.2"
 tracing = { version = "^0.1", features = ["std"] }
 tokio = { version = "1", features = ["time", "net", "io-util"] }
 drmem-api = { path = "../../drmem-api", version = "0.1" }
-drmem-config = { path = "../../drmem-config", version = "0.1" }
 reqwest = { default-features = false, version = "0.11" }
 weather-underground = "0.1"
