--- drmem-config/Cargo.toml
@@ -1,30 +1,0 @@
-[package]
-name = "drmem-config"
-version = "0.1.0"
-authors = ["Rich Neswold <rich.neswold@gmail.com>"]
-edition = "2021"
-description = "Configuration crate for DrMem control system"
-homepage = "https://github.com/DrMemCS/drmem"
-repository = "https://github.com/DrMemCS/drmem"
-license = "MIT"
-categories = ["embedded"]
-keywords = ["control-system", "automation"]
-
-[badges]
-maintenance = { status = "actively-developed" }
-
-[dependencies]
-async-trait = "0.1"
-clap = { version = "2", default-features = false }
-toml = "0.5.6"
-serde = "1.0.115"
-serde_derive = "1.0.115"
-tracing-subscriber = "^0.2"
-tracing-futures = "^0.2"
-tracing = { version = "^0.1", features = ["std"] }
-tokio = { version = "1", features = ["rt-multi-thread", "macros", "time", "fs"] }
-drmem-api = { path = "../drmem-api", version = "0.1" }
-
-[features]
-redis-backend = []
-graphql = []
