--- drmem-config/README.md
@@ -1,4 +1,0 @@
-# drmem-config
-
-This crate adds configuration file support for the [DrMem control
-system](https://github.com/DrMemCS/drmem).
