+++ drmemd/src/graphql/mod.rs
@@ -459,9 +459,7 @@ fn schema() -> Schema {
 }
 
 pub fn server(
-    node_name: &str,
-    addr: &std::net::SocketAddr,
-    db: crate::driver::DriverDb,
+    node_name: &str, addr: &std::net::SocketAddr, db: crate::driver::DriverDb,
     cchan: client::RequestChan,
 ) -> impl Future<Output = ()> {
     const FULL_QUERY_PATH: &str = "/query";
@@ -550,7 +548,7 @@ pub fn server(
 
     let service = resp.register(
         "_drmem._tcp".into(),
-	format!("DrMem-{}", node_name),
+        format!("DrMem-{}", node_name),
         addr.port(),
         &[
             &format!("queries={}", FULL_QUERY_PATH),
